@@ -1,15 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="AspNetCore.Identity.MongoDbCore" Version="1.1.1" />
-    <PackageReference Include="AWSSDK.CognitoIdentityProvider" Version="3.3.102.34" />
-    <PackageReference Include="Bogus" Version="28.0.2" />
-    <PackageReference Include="Hangfire.Core" Version="1.7.5" />
-    <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="2.2.0" />
+    <PackageReference Include="AWSSDK.CognitoIdentityProvider" Version="3.3.105.2" />
+    <PackageReference Include="Bogus" Version="28.4.1" />
+    <PackageReference Include="Hangfire.Core" Version="1.7.7" />
+    <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="3.0.0" />
     <PackageReference Include="RestSharp" Version="106.6.10" />
   </ItemGroup>
 
