@@ -1,14 +1,12 @@
 ﻿using QuarklessContexts.Models;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Quarkless.Services.Interfaces.Actions
 {
 	public class AccountCheckerActionOptions : IActionOptions
 	{
 		public DateTimeOffset ExecutionTime { get; set; }
-		public static Range TimeFrameSeconds { get; set; } = new Range(1800, 2000);
+		public static XRange TimeFrameSeconds { get; set; } = new XRange(1800, 2000);
 
 	}
 }
