@@ -7,7 +7,7 @@ namespace Quarkless.Services.Interfaces.Actions
 	{
 		public int ImageFetchLimit { get; set; } = 20;
 		public DateTimeOffset ExecutionTime { get; set; }
-		public static Range TimeFrameSeconds { get; set; } = new Range(1500, 2000);
+		public static XRange TimeFrameSeconds { get; set; } = new XRange(1500, 2000);
 
 		public PostActionOptions(DateTimeOffset executionTime)
 		{
