@@ -6,7 +6,7 @@ namespace Quarkless.Services.Interfaces.Actions
 	public class UnfollowActionOptions : IActionOptions
 	{
 		public DateTimeOffset ExecutionTime { get; set; }
-		public static Range TimeFrameSeconds { get; set; } = new Range(45, 75);
+		public static XRange TimeFrameSeconds { get; set; } = new XRange(45, 75);
 
 		public UnfollowActionOptions(DateTimeOffset executionTime)
 		{
