@@ -3,7 +3,8 @@ using System.IO;
 using Hangfire;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Quarkless.Common;
+using Quarkless.Common.Clients;
+using QuarklessContexts.Models.SecurityLayerModels;
 using QuarklessLogic.QueueLogic.Jobs.Filters;
 using StackExchange.Redis;
 
@@ -11,17 +12,45 @@ namespace Quarkless.Tasks
 {
 	class Program
 	{
-		static void Main(string[] args)
+		private const string CLIENT_SECTION = "Client";
+		private static IConfiguration MakeConfigurationBuilder()
 		{
-			var services = new ServiceCollection();
-			var settingPath = Path.GetFullPath(Path.Combine(Accessors.BasePath, @"Quarkless"));
-			IConfiguration configuration = new ConfigurationBuilder().SetBasePath(settingPath).AddJsonFile("appsettings.json").Build();
-			var accessors = new Accessors(configuration);
-			var connectionMultiplexer = ConnectionMultiplexer.Connect(accessors.RedisConnectionString);
+			return new ConfigurationBuilder()
+				.SetBasePath(Directory.GetCurrentDirectory().Split("bin")[0])
+				.AddJsonFile("appsettings.json").Build();
+		}
+		private static void InitialiseClientServices()
+		{
+			var cIn = new ClientRequester();
+			if (!cIn.TryConnect().GetAwaiter().GetResult())
+				throw new Exception("Invalid Client");
 
-			services.AddHangFrameworkServices(accessors);
+			var caller = MakeConfigurationBuilder().GetSection(CLIENT_SECTION).Get<AvailableClient>();
 
-			GlobalConfiguration.Configuration.UseRedisStorage(connectionMultiplexer, new Hangfire.Redis.RedisStorageOptions{
+			var validate = cIn.Send(new InitCommandArgs
+			{
+				Client = caller,
+				CommandName = "Validate Client"
+			});
+			if (!(bool)validate)
+				throw new Exception("Could not validated");
+
+			var services = (IServiceCollection) cIn.Send(new BuildCommandArgs()
+			{
+				CommandName = "Build Services",
+				ServiceTypes = new[]
+				{
+					ServiceTypes.AddHangFrameworkServices
+				}
+			});
+			var endPoints = (EndPoints) cIn.Send(new GetPublicEndpointCommandArgs
+			{
+				CommandName = "Get Public Endpoints",
+			});
+			var connectionMultiplexer = ConnectionMultiplexer.Connect(endPoints.RedisCon);
+
+			GlobalConfiguration.Configuration.UseRedisStorage(connectionMultiplexer, new Hangfire.Redis.RedisStorageOptions
+			{
 				Prefix = "Timeline",
 				SucceededListSize = 5000,
 				DeletedListSize = 1000,
@@ -29,14 +58,16 @@ namespace Quarkless.Tasks
 				InvisibilityTimeout = TimeSpan.FromMinutes(30),
 				UseTransactions = true
 			}).WithJobExpirationTimeout(TimeSpan.FromDays(1));
-
 			GlobalConfiguration.Configuration.UseActivator(new WorkerActivator(services.BuildServiceProvider(false)));
 			GlobalConfiguration.Configuration.UseSerializerSettings(new Newtonsoft.Json.JsonSerializerSettings()
 			{
 				ReferenceLoopHandling = Newtonsoft.Json.ReferenceLoopHandling.Ignore
-			}); 	
+			});
 			GlobalJobFilters.Filters.Add(new ProlongExpirationTimeAttribute());
+		}
 
+		static void Main(string[] args)
+		{
 			#region Mongo Storage
 			//GlobalConfiguration.Configuration.UseMongoStorage(_accessors.ConnectionString, _accessors.SchedulerDatabase, new MongoStorageOptions()
 			//{
@@ -51,14 +82,13 @@ namespace Quarkless.Tasks
 			//});
 			#endregion
 
-
-			BackgroundJobServerOptions jobServerOptions = new BackgroundJobServerOptions
+			InitialiseClientServices();
+			var jobServerOptions = new BackgroundJobServerOptions
 			{
 				WorkerCount = Environment.ProcessorCount * 5,
 				ServerName = string.Format("TASK_{0}.{1}",Environment.MachineName,Guid.NewGuid().ToString()),
 				//Activator = new WorkerActivator(services.BuildServiceProvider(false)),
 			};
-
 			using (var server = new BackgroundJobServer(jobServerOptions))
 			{
 				Console.WriteLine("hangfire server started");
