@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Internal;
 using System;
 using System.Diagnostics;
 using System.IO;
@@ -25,7 +24,9 @@ namespace Quarkless.Filters
 		private readonly SecurityHeadersPolicy _policy;
 		#endregion
 
-		public RequestResponseLoggingMiddleware(RequestDelegate next, SecurityHeadersPolicy securityHeadersPolicy, IAPILogCache apiLogCache,
+		public RequestResponseLoggingMiddleware(RequestDelegate next, 
+			SecurityHeadersPolicy securityHeadersPolicy, 
+			IAPILogCache apiLogCache,
 			IOptions<MaxConcurrentRequestsOptions> options)
 		{
 			_concurrentRequestsCount = 0;
@@ -159,8 +160,7 @@ namespace Quarkless.Filters
 			var body = request.Body;
 
 			//This line allows us to set the reader for the request back at the beginning of its stream.
-			request.EnableRewind();
-
+			request.EnableBuffering();
 			//We now need to read the request stream.  First, we create a new byte[] with the same length as the request stream...
 			var buffer = new byte[Convert.ToInt32(request.ContentLength)];
 
