@@ -1,7 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
 using AspNetCoreRateLimit;
 using Microsoft.AspNetCore;
@@ -17,22 +13,19 @@ namespace Quarkless
     {
         public static async Task Main(string[] args)
         {
-            var webHost = BuildWebHost(args);
+	        var webHost = BuildWebHost(args);
             using (var scope = webHost.Services.CreateScope())
             {
-	            // get the IpPolicyStore instance
 	            var ipPolicyStore = scope.ServiceProvider.GetRequiredService<IIpPolicyStore>();
-
-	            // seed IP data from appsettings
 	            await ipPolicyStore.SeedAsync();
             }
 
             await webHost.RunAsync();
         }
 
-        public static IWebHost BuildWebHost(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
-                .Build();
-    }
+		public static IWebHost BuildWebHost(string[] args) =>
+			WebHost.CreateDefaultBuilder(args)
+				.UseStartup<Startup>()
+				.Build();
+	}
 }
