@@ -20,16 +20,16 @@ namespace QuarklessLogic.ServicesLogic.ContentSearch
 		Task<List<UserResponse<string>>> SearchInstagramMediaLikers(string mediaId);
 		Task<Media> SearchRecentLocationMediaDetailInstagram(Location location, int limit);
 		Task<Media> SearchTopLocationMediaDetailInstagram(Location location, int limit);
-		Task<Media> SearchMediaDetailInstagram(List<string> topics, int limit, bool isRecent = false);
-		Task<Media> SearchMediaInstagram(List<string> topics, InstaMediaType mediaType, int limit);
+		Task<Media> SearchMediaDetailInstagram(IEnumerable<string> topics, int limit, bool isRecent = false);
+		Task<Media> SearchMediaInstagram(IEnumerable<string> topics, InstaMediaType mediaType, int limit);
 		Task<Media> SearchMediaUser(string username = null, int limit = 1);
 		Task<Media> SearchUserFeedMediaDetailInstagram(string[] seenMedias = null, bool requestRefresh = false, int limit = 1);
 		Task<Media> SearchUsersMediaDetailInstagram(string userName, int limit);
 		SearchResponse<Media> SearchViaGoogle(SearchImageModel searchImageQuery);
 		SearchResponse<Media> SearchViaYandex(YandexSearchQuery yandexSearchQuery, int limit);
-		SearchResponse<Media> SearchYandexSimilarSafeMode(List<GroupImagesAlike> imagesAlikes, int limit);
-		SearchResponse<Media> SearchViaYandexBySimilarImages(List<GroupImagesAlike> imagesSimilarUrls, int limit, int offset = 0);
-		SearchResponse<Media> SearchSimilarImagesViaGoogle(List<GroupImagesAlike> groupImages, int limit, int offset = 0);
+		SearchResponse<Media> SearchYandexSimilarSafeMode(IEnumerable<GroupImagesAlike> imagesAlikes, int limit);
+		SearchResponse<Media> SearchViaYandexBySimilarImages(IEnumerable<GroupImagesAlike> imagesSimilarUrls, int limit, int offset = 0);
+		SearchResponse<Media> SearchSimilarImagesViaGoogle(IEnumerable<GroupImagesAlike> groupImages, int limit, int offset = 0);
 		Task<IEnumerable<TopicCategories>> GetBusinessCategories();
 		void ChangeUser(IAPIClientContainer newUser);
 		Task<InstaDirectInboxContainer> SearchUserInbox(int limit = 1);
